PetPanel - Created 4 separate components for every action that get and sends action data. Changed screen names (added word Screen). Added folders to components to group buttons, image picker and PetScreen components.
