+++ React_Native/petnanny/components/PetListItem.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Image, Pressable, StyleSheet, Text, View} from 'react-native';
+import {Pressable, StyleSheet, Text, View} from 'react-native';
 import {Colors} from '../constants/colors';
 import {Card} from './Card';
 import {PetPhoto} from './PetPhoto';
