+++ React_Native/petnanny/screens/RegisterScreen.js
@@ -1,16 +1,14 @@
 import React from 'react';
-import {Alert, StyleSheet, TextInput, View} from 'react-native';
+import {Alert, StyleSheet, TextInput, View, Text} from 'react-native';
 import {Logo} from '../components/Logo';
 import {Colors} from '../constants/colors';
-import {Button} from '../components/Button';
+import {Button} from '../components/buttons/Button';
 import {useState} from 'react';
-import {Title} from '../components/Title';
 import auth from '@react-native-firebase/auth';
 
-export const Login = () => {
+export const RegisterScreen = () => {
   const [emailAddress, setEmailAdress] = useState('');
   const [password, setPassword] = useState('');
-  const [userName, setUserName] = useState('');
 
   const emailHandler = enteredText => {
     setEmailAdress(enteredText);
@@ -19,31 +17,44 @@ export const Login = () => {
     setPassword(enteredText);
   };
 
-  // const validateEmail = emailAddress => {
-  //   var re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  //   return re.test(emailAddress);
-  // };
-
-  // console.log(validateEmail('anystring@anystring.anystring'));
-
   const onPressHandler = () => {
-    auth()
-      .signInWithEmailAndPassword(emailAddress, password)
-      .then(() => {
-        setEmailAdress('');
-        setPassword('');
-      })
-      .catch(error => {
-        if (error.code === 'auth/invalid-password') {
-          Alert.alert('Invalid password!', 'Please, try again.');
-        }
+    if (
+      password.match(/[a-z]/g) &&
+      password.match(/[A-Z]/g) &&
+      password.match(/[0-9]/g) &&
+      password.match(/[^a-zA-Z\d]/g) &&
+      password.length >= 8
+    ) {
+      auth()
+        .createUserWithEmailAndPassword(emailAddress, password)
+        .then(() => {
+          setEmailAdress('');
+          setPassword('');
+        })
+        .catch(error => {
+          if (error.code === 'auth/email-already-in-use') {
+            Alert.alert(
+              'That email address is already in use!',
+              'Please, use a different email.',
+            );
+          }
 
-        if (error.code === 'auth/invalid-email') {
-          Alert.alert('Invalid email!', 'Please, try again.');
-        }
+          if (error.code === 'auth/invalid-email') {
+            Alert.alert(
+              'That email address is invalid!',
+              'Please, use a different email.',
+            );
+          }
 
-        console.error(error);
-      });
+          console.error(error);
+        });
+    } else {
+      Alert.alert(
+        'Invalid password!',
+        'The password should contain at least 8 characters, including at least one uppercase letter and one lowercase letter, one special character, and one number.',
+        [{text: 'OK', style: 'destructive', onPress: setPassword('')}],
+      );
+    }
   };
 
   return (
@@ -74,7 +85,7 @@ export const Login = () => {
         />
       </View>
       <View style={styles.buttonContainer}>
-        <Button onPress={onPressHandler}>Login</Button>
+        <Button onPress={onPressHandler}>Register</Button>
       </View>
     </View>
   );
